third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/simple-cross-origin-progress-events-expected.txt

CONSOLE ERROR: Failed to load http://localhost:8000/xmlhttprequest/resources/cross-site-progress-events.cgi: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Failed to load http://localhost:8000/xmlhttprequest/resources/cross-site-progress-events.cgi: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
Test that upload progress events are not dispatched for simple cross-origin requests (i.e. if the listener is set after calling send(), and there are no other reasons to make a preflight request).

Test 1: The URL is allowed for cross-origin requests
onprogress
onload
Response length: 8388608

Test 2: The URL is not allowed for cross-origin requests
onprogress
onerror (expected)
Response length: 0

Test 3: The URL is not allowed for cross-origin requests and at least one upload event is listened for before doing the send.
upload.onloadstart
upload.onprogress
upload.onerror (expected)
onprogress
onerror (expected)
Response length: 0
